Extract Outcome class in separate file
