+++ data/src/main/java/com/aconno/sensorics/data/repository/action/ActionsRepositoryImpl.kt
@@ -4,7 +4,7 @@ import com.aconno.sensorics.domain.actions.Action
 import com.aconno.sensorics.domain.actions.ActionsRepository
 import com.aconno.sensorics.domain.actions.GeneralAction
 import com.aconno.sensorics.domain.ifttt.LimitCondition
-import com.aconno.sensorics.domain.ifttt.outcome.Outcome
+import com.aconno.sensorics.domain.actions.outcomes.Outcome
 import com.aconno.sensorics.domain.model.Device
 import io.reactivex.Single
 
@@ -61,7 +61,10 @@ class ActionsRepositoryImpl(
             Pair(Outcome.PHONE_NUMBER, actionEntity.phoneNumber)
         )
 
-        val outcome = Outcome(parameters, actionEntity.outcomeType)
+        val outcome = Outcome(
+            parameters,
+            actionEntity.outcomeType
+        )
 
         val device = Device(
             actionEntity.deviceName,
