+++ domain/src/main/java/com/aconno/sensorics/domain/actions/GeneralAction.kt
@@ -1,7 +1,7 @@
 package com.aconno.sensorics.domain.actions
 
 import com.aconno.sensorics.domain.ifttt.Condition
-import com.aconno.sensorics.domain.ifttt.outcome.Outcome
+import com.aconno.sensorics.domain.actions.outcomes.Outcome
 import com.aconno.sensorics.domain.model.Device
 
 class GeneralAction(
