+++ domain/src/main/java/com/aconno/sensorics/domain/ifttt/outcome/RunOutcomeUseCase.kt
@@ -2,6 +2,7 @@ package com.aconno.sensorics.domain.ifttt.outcome
 
 import com.aconno.sensorics.domain.SmsSender
 import com.aconno.sensorics.domain.Vibrator
+import com.aconno.sensorics.domain.actions.outcomes.Outcome
 import com.aconno.sensorics.domain.ifttt.NotificationDisplay
 import com.aconno.sensorics.domain.ifttt.TextToSpeechPlayer
 import com.aconno.sensorics.domain.interactor.type.CompletableUseCaseWithParameter
@@ -36,29 +37,6 @@ class OutcomeExecutorSelector(
             Outcome.OUTCOME_TYPE_VIBRATION -> vibrationOutcomeExecutor
             else -> throw IllegalArgumentException("Invalid Outcome type.")
         }
-    }
-}
-
-class Outcome(val parameters: Map<String, String>, val type: Int) {
-
-    override fun toString(): String {
-        return when (type) {
-            0 -> "Notification"
-            1 -> "SMS"
-            2 -> "Text to Speech"
-            3 -> "Vibration"
-            else -> throw IllegalArgumentException("Outcome type not valid: $type")
-        }
-    }
-
-    companion object {
-        const val OUTCOME_TYPE_NOTIFICATION = 0
-        const val OUTCOME_TYPE_SMS = 1
-        const val OUTCOME_TYPE_TEXT_TO_SPEECH = 2
-        const val OUTCOME_TYPE_VIBRATION = 3
-
-        const val TEXT_MESSAGE = "textMessage"
-        const val PHONE_NUMBER = "phoneNumber"
     }
 }
