+++ domain/src/main/java/com/aconno/sensorics/domain/interactor/ifttt/InputToOutcomesUseCase.kt
@@ -3,7 +3,7 @@ package com.aconno.sensorics.domain.interactor.ifttt
 import com.aconno.sensorics.domain.actions.Action
 import com.aconno.sensorics.domain.actions.ActionsRepository
 import com.aconno.sensorics.domain.ifttt.Input
-import com.aconno.sensorics.domain.ifttt.outcome.Outcome
+import com.aconno.sensorics.domain.actions.outcomes.Outcome
 import com.aconno.sensorics.domain.interactor.type.SingleUseCaseWithParameter
 import io.reactivex.Observable
 import io.reactivex.Single
